Bhargav/requirements.txt
streamlit
pandas
scikit-learn
nltk
wordcloud
pypdf
python-docx
reportlab